Remove ability to disable the client at runtime
One of the requirements we would need to maintain is that the profiler
is started and shutdown in the same thread in all cases. This seems like
a hard problem, so lets just not solve it for now.

I don't think people will have significant issues with the profiler just
being there until the process terminates (not to mention that kernel is
probably much better at cleaning up everything than we are anyway).