Set default callstack_depth to 0
Collecting callstacks is quite expensive and can make each
instrumentation point ~100 times slower (200ns -> 20µs). Originally
callstack collection was enabled by default in a hope to provide a great
out-of-the-box experience by just enabling the layer or inserting some
instrumentation points. In retrospective that is a poor default.

Compared to figuring out why the performance is suffering, it seems like
it would be _much_ easier for user to discover how to enable collection
of stack frames when they find themselves wanting to see some in their
profiles. It also gives us an opportunity to visibly document the
pitfalls of enabling collection of the stack frames.