ci: update `bindgen` installation
This crate has changed its delivery mechanism recently, which seems to
have broken us. I’m of a mind that perhaps we ought to be locking the
version of bindgen, but it is also plausible that the changes in this
area are going to be rare enough for it to not to become a maintenance
burden.