Explicitly specify which kind of library to link
This is quite important, for a couple of reasons. First, we must
manually link the dependencies (`libdl`/`libc`/`user32`) when linking to
a static library, as static library archives don’t carry such dependnecy
information by themselves on many platforms.

And second, it just a good style to explicitly specify the intent so
that rustc doesn’t pick at random depending on exact configuration of
the user’s system or whatever.