Simplify the GHAs
Primary motivation is to remove the dependency on now-unmaintained
actions-rs. The resulting code is tad a bit cleaner either way.

We also bump the version we test with to 1.56.0 for that’s the version
required by `once_cell`. That said, there isn’t any change in our code
that would actually bump MSRV past 1.53.0, but 1.56 is also old
enough that I have no qualms over bumping to that MSRV.