Allow specifying the C++ edition
In certain build environments *cough ChromeOS cough* the C++ edition may
be hamstrung. For example, chrono may not be available - see
https://chromium.googlesource.com/chromium/src/+/HEAD/styleguide/c++/c++-features.md#chrono_banned.

To workaround this, allow an environment variable to specify the
edition. For example: `TRACY_CLIENT_CPP_EDITION=gnu++14 cargo build`

Note that since compilers seem to be resilient to badly formed editions,
the implementation itself will check for correctness.