config: perhaps drop DynamicConfig too?
In practice I have found that, whenever a library provides an ability to
configure through a trait generic and then provides a basic runtime
cofniguration convenience-type, the convenience type ends up being
dropped in favour of a manual implementation pretty quickly.

So it seems like a not-terrible-idea to just push users to the trait
immediately? And we don’t need to figure a good design for a
compile-time configuration type…

Might still go back on this decision, idk.