ci: test libraries individually
This verifies that the libraries can be built on their own. Defects
in a test that builds them all together can be hidden in case a
workspace crate accidentally enables a feature (`--features=enable`)
which the CI job ought to check builds with it disabled (e.g. examples
crate depending on `tracy-client` with its default features enabled.)