fix stacktraces on Windows again
It turns out that formatting arguments get evaluated lazily, which means the call to `std::fmt::Debug` of the `std::backtrace::Backtrace` could be optimized out.

Rust 1.87.0 (which upgraded to LLVM 20) actually started removing that call.

This is less likely to break, because it creates a `String` using the backtrace and passes that `String` to `std::hint::black_box`. This isn't guaranteed to work, but should be good enough.

Alternatives considered:
- Calling `backtrace::dbghelp::init` directly. Not possible because the `dbghelp` module is private, and it would require a dependency on `backtrace`.
- Manually resolving something through the `backtrace` crate. Probably the sanest alternative, but would require a dependency on `backtrace`.
- Initializing `dbghelp.dll` ourselves. Significant amount of (unsafe) code to write/vendor. And it could easily drift from what stdlib/backtrace are doing.