fix stacktraces on Windows again (#142)
It turns out that formatting arguments get evaluated lazily, which means the call to `std::fmt::Debug` of the `std::backtrace::Backtrace` could be optimized out.

Rust 1.87.0 (which upgraded to LLVM 20) actually started removing that call.

This stops using the workaround of resolving a backtrace from the standard library to force initialization of the symbol helper.
Instead, we do the initialization ourselves and add the directory of all loaded modules to the symbol search path. This matches the behavior of the standard library/`backtrace-rs`.